refactor: restructured AwaitingShipmentCard & AwaitingShipmentTable, into app/client/component
chore: added awaitingshipment.md documentation
